clean up /reorganize table package
