+++ packages/Charts/src/Histogram/Histogram.tsx
@@ -171,7 +171,7 @@ export const ThresholdSelectHistogram = ({
                 ...opts,
             });
         }
-    }, []);
+    }, [data]);
 
     return (
         <>
