+++ packages/Table/src/Cells/CellRenderers/TextRenderer.tsx
@@ -1,7 +1,7 @@
-import { DEFAULT_NA_VALUE } from "../Cell";
+import { DEFAULT_NA_VALUE } from "../../types";
 import React from "react";
 import { _isNA } from "@niagads/common";
-import styles from "../styles/cell.module.css";
+import styles from "./cell.module.css";
 
 export type LinkTarget = "_blank" | "_self" | "_parent" | "_top";
 
@@ -10,5 +10,5 @@ export interface TextRenderer<T> {
 }
 
 export const renderNullValue = (value: string = DEFAULT_NA_VALUE) => {
-    return <span className={styles.nullValue}>{value}</span>;
+    return <span className={styles["null-value"]}>{value}</span>;
 };
