+++ packages/Table/src/Cells/utils.tsx
@@ -1,100 +1,13 @@
-import { Badge, BadgeIconType, BooleanBadge } from "./CellRenderers/Badge";
-import {
-    BasicType,
-    Expand,
-    Modify,
-    NAString,
-    TypeMapper,
-    _deepCopy,
-    _get,
-    _hasOwnProperty,
-    _isJSON,
-    _isNA,
-    _isNull,
-} from "@niagads/common";
-import { Link, LinkList } from "./CellRenderers/Link";
-import React, { CSSProperties } from "react";
-import { Text, TextList } from "./CellRenderers/BasicText";
+import { Badge, BooleanBadge, Float, Link, LinkList, PercentageBar, Text, TextList } from "./CellRenderers";
+import { BasicType, _deepCopy, _get, _hasOwnProperty, _isJSON, _isNA, _isNull } from "@niagads/common";
+import { BooleanCell, CELL_TYPES, Cell, CellType, FloatCell, PValueCell, TableCell } from "./types";
+import { DEFAULT_NA_VALUE, TableColumn } from "../types";
 
-import { Float } from "./CellRenderers/Number";
-import { LinkTarget } from "./CellRenderers";
-import { PercentageBar } from "./CellRenderers/SparkChart";
-import { TableColumn } from "./types";
-
-export const DEFAULT_NA_VALUE = "n/a";
-
-export type GenericCell = BasicType | Record<string, any> | null;
-
-export type AbstractCell = {
-    type: "abstract";
-    value: BasicType | null;
-    rowId: number;
-    columnId: number;
-    nullValue?: BasicType | null; // if not set, treats null as NA
-    naValue?: NAString; // (internal) value to assign to NAs for consistency, defaults to `NA`
-    className?: string;
-    style?: CSSProperties;
-};
-
-export type IntegerCell = Expand<Modify<AbstractCell, { type: "integer"; value: number | null }>>;
-
-export type FloatCell = Expand<Modify<AbstractCell, { type: "float"; value: number | null; precision?: number }>>;
-
-export type PValueCell = Expand<Modify<FloatCell, { type: "pvalue" }>>;
-
-export type TextCell = Expand<Modify<AbstractCell, { type: "text"; truncateTo?: number; info?: string }>>;
-
-export type TextListCell = Expand<Modify<AbstractCell, { type: "text_list"; value: string; items: TextCell[] }>>;
-
-export type BadgeCell = Expand<
-    Modify<
-        TextCell,
-        {
-            type: "badge";
-            icon?: BadgeIconType;
-            iconOnly?: boolean;
-        }
-    >
->;
-
-export type BooleanCell = Expand<
-    Modify<
-        BadgeCell,
-        {
-            type: "boolean";
-            value: boolean | null;
-            displayText?: BasicType; // what value to display (e.g., TRUE, Yes, Y, Coding, FALSE, N, No, etc)
-        }
-    >
->;
-
-export type LinkCell = Expand<Modify<AbstractCell, { type: "link"; url: string; info?: string; target?: LinkTarget }>>;
-
-export type LinkListCell = Expand<Modify<AbstractCell, { type: "link_list"; value: string; items: LinkCell[] }>>;
-
-export type PercentageBarCell = Expand<Modify<FloatCell, { type: "percentage_bar" }>>;
-
-export type Cell =
-    | PercentageBarCell
-    | FloatCell
-    | IntegerCell
-    | PValueCell
-    | AbstractCell
-    | TextCell
-    | TextListCell
-    | BadgeCell
-    | BooleanCell
-    | LinkCell
-    | LinkListCell;
-
-// create CellType which is a list string keys corresponding to allowable "types" of cells
-type CellTypeMapper = TypeMapper<Cell>;
-export type CellType = keyof CellTypeMapper;
+import React from "react";
 
 // validates cell type specified at runtime or by user is valid
 // if cell type is undefined, returns "abstract"
 // this does not include LinkList & TextList b/c those are internal cell types
-const CELL_TYPES = ["percentage_bar", "float", "text", "abstract", "integer", "pvalue", "badge", "boolean", "link"];
 
 export const validateCellType = (ctype: string | undefined): CellType => {
     if (ctype === undefined) {
@@ -154,15 +67,15 @@ export const getCellValue = (cellProps: Cell | Cell[]): any => {
     }
 };
 
-const __resolveLinkCell = (cell: GenericCell): GenericCell => {
+const __resolveLinkCell = (cell: TableCell): TableCell => {
     const value = _get("value", cell);
     const url = _get("url", cell);
     Object.assign(cell as any, { value: value ? value : url, type: "link" });
     return cell;
 };
 
-const __resolveListCell = (cells: GenericCell[]) => {
-    const values = cells.map((c: GenericCell) => _get("value", c));
+const __resolveListCell = (cells: TableCell[]) => {
+    const values = cells.map((c: TableCell) => _get("value", c));
     const value = values.join(" // ");
 
     return { type: "abstract", value: value, items: cells };
@@ -174,12 +87,12 @@ const __resolveListCell = (cells: GenericCell[]) => {
 // 1. makes sure user specified a cell type to the parent column if cell value is an object/JSON
 
 export const resolveCell = (
-    cell: GenericCell | GenericCell[],
+    cell: TableCell | TableCell[],
     column: TableColumn,
     rowId: number
-): GenericCell | GenericCell[] => {
+): TableCell | TableCell[] => {
     let resolvedCellType = _get("type", column, "abstract");
-    let resolvedCell: GenericCell = {};
+    let resolvedCell: TableCell = {};
 
     if (Array.isArray(cell)) {
         if (resolvedCellType == "abstract") {
@@ -190,7 +103,7 @@ export const resolveCell = (
             );
         }
 
-        const cellList = cell.map((c: GenericCell) => resolveCell(c, column, rowId) as GenericCell);
+        const cellList = cell.map((c: TableCell) => resolveCell(c, column, rowId) as TableCell);
         if (resolvedCellType == "text") {
             resolvedCell = __resolveListCell(cellList);
             resolvedCellType = "text_list";
@@ -218,12 +131,6 @@ export const resolveCell = (
             } else {
                 resolvedCellType = "text";
             }
-            /* console.warn(
-                "`type` must be specified in the column defintion to correctly interpret structured values; assuming `" +
-                    resolvedCellType +
-                    "` cell: " +
-                    JSON.stringify(cell)
-            ); */
         }
 
         if (_get("value", resolvedCell) == null) {
