+++ packages/Table/src/ControlElements/Columns/ColumnControls.tsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useId, useState } from "react";
+import React, { useState } from "react";
+import { TableColumn, TableRow } from "@/src/types";
 import { TextInput, Toggle, ToggleGroup } from "@niagads/ui";
 
 import { ActionMenu } from "@niagads/ui/client";
 import { Column } from "@tanstack/react-table";
 import { Columns3 } from "lucide-react";
-import { TableRow, TableColumn } from "../types";
-import styles from "../styles/controls.module.css";
+import styles from "./ColumnControls.module.css";
 
 interface ColumnControlsProps {
     columns: Column<TableRow, unknown>[];
