+++ packages/Table/src/ControlElements/controls.module.css
@@ -37,12 +37,3 @@
     gap: 0.5rem;
     margin: 0.5rem;
 }
-
-.column-controls-container {
-    width: 200px;
-}
-
-.column-controls-container,
-.column-controls-toggle-container {
-    padding: 0.5rem;
-}
