+++ packages/Table/src/ControlElements/index.ts
@@ -1,5 +1,5 @@
 export * from "./PaginationControls";
 export * from "./TableToolbar";
-export * from "./ColumnControls";
+export * from "./Columns/ColumnControls";
 export * from "./RowSelectionControls";
 export * from "./TableExportControls";
