--- packages/Table/src/styles/ui.css
@@ -1,86 +1,0 @@
-/* UI ELEMENTS */
-
-.icon-disabled {
-    stroke: #d1d5db; /* Tailwind slate-300 */
-}
-
-.info-bubble {
-    color: transparent;
-    width: 1rem;
-    height: 1rem;
-    stroke: var(--color-primary, #2563eb); /* Tailwind primary */
-}
-
-/* FIXME: for formatting icon in a button icon */
-.icon-button {
-    width: 1rem;
-    height: 1rem;
-    stroke-width: 4;
-}
-
-.icon {
-    width: 1rem;
-    height: 1rem;
-}
-
-.icon.clickable {
-    cursor: pointer;
-}
-
-.pill {
-    display: flex;
-    align-items: center;
-    border-style: solid;
-    border-width: thin;
-    border-radius: 9999px;
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    margin-right: 5px;
-    margin-left: 5px;
-}
-
-.stroke-1 {
-    stroke-width: 1;
-}
-
-.stroke-black {
-    stroke: #000;
-}
-
-.stroke-white {
-    stroke: white;
-}
-
-.hidden {
-    display: none;
-}
-
-.dropdown:focus-within .dropdown-menu {
-    display: block;
-}
-
-.dropdown-menu-button {
-    z-index: 50;
-    position: absolute;
-    right: 0;
-    width: 14rem; /* 56 x 0.25rem */
-    margin-top: 0.5rem;
-    transform-origin: top right;
-    background: #fff;
-    border: 1px solid #e5e7eb; /* Tailwind gray-200 */
-    border-radius: 0.375rem; /* Tailwind rounded-md */
-    box-shadow:
-        0 10px 15px -3px rgba(0, 0, 0, 0.1),
-        0 4px 6px -4px rgba(0, 0, 0, 0.1); /* Tailwind shadow-lg */
-    outline: none;
-    /* divide-y and divide-gray-100 handled by child elements if needed */
-}
-
-.dropdown-menu-item-container {
-    padding-left: 1rem;
-    padding-right: 1rem;
-    padding-top: 0.75rem;
-    padding-bottom: 0.75rem;
-}
